requirements.txt
elasticsearch
aiohttp
pyyaml